The "Killed By" Complete Update.
Base enemies pull from an array of common names and prepend that to the given enemy name to create a unique name that is then set to that instance of an enemyName, and when they kill the Player that name is passed to the Lose Menu to display who killed the player. The Bosses are already uniquely named so we just pass that on when they kill the Player. Adjusted the Zombie height a tad.