FIXED - Enemies just run into walls while roaming or fleeing
Ended up sorting thru quite a lot of stuff with base enemies. Moved the fire particle effect into the model's bodies so the animation will now move/rotate with the model better. Set AI roaming to move then pause correctly, and set canSeePlayer so that when the enemy loses sight of the player they move to the last known position of the player before going into roaming mode, this is way more effective for tracking the Player around corners. Corrected an issue in the death animation that was causing the weird falling positions that resulted in a lot of feet in the air and not laying prone, the setting was the Root Transform Position (Y) Based On set to "Original", now changed to "Feet", so the enemy drops to their feet.